:lipstick: moved bot creation to dialog
