+++ app/frontend/src/pages/layout/Layout.module.css
@@ -2,6 +2,7 @@
     min-height: 100vh;
     display: flex;
     flex-direction: column;
+    width: 100%;
 }
 
 .headerContainer {
@@ -39,6 +40,8 @@
     flex-wrap: wrap;
     justify-items: center;
     justify-content: center;
+    align-self: center;
+    justify-self: center;
     width: 60%;
 }
